+++ src/monk.js
@@ -25,9 +25,6 @@ async function flurryOfBlows(actor) {
         return;
     }
 
-    const DamageRoll = CONFIG.Dice.rolls.find( r => r.name === "DamageRoll" );
-    const critRule = game.settings.get("pf2e", "critRule");
-
     const weapons = flurryOfBlowsWeapons(actor)
     if (weapons.length === 0) {
         ui.notifications.warn(`${actor.name} doesn't have correct weapon'`);
@@ -39,7 +36,7 @@ async function flurryOfBlows(actor) {
         weaponOptions += `<option value=${w.item.id}>${w.item.name}</option>`
     }
 
-    const { currentWeapon, map, dos} = await Dialog.wait({
+    const { currentWeapon, map } = await Dialog.wait({
         title:"Flurry of Blows",
         content: `
             <div class="row-flurry"><div class="column-flurry first-flurry"><h3>First Attack</h3><select id="fob1" autofocus>
@@ -74,95 +71,15 @@ async function flurryOfBlows(actor) {
         default: "ok"
     });
 
+    if ( currentWeapon === undefined || map === undefined ) { return; }
     const map2 = map === 2 ? map : map + 1;
-    if ( currentWeapon === undefined ) { return; }
 
     const primary = weapons.find( w => w.item.id === currentWeapon[0] );
     const secondary = weapons.find( w => w.item.id === currentWeapon[1] );
 
     const options = actorFeat(actor, "stunning-fist") ? ["stunning-fist"] : [];
 
-    const damages = [];
-    function PD(cm) {
-        if ( cm.user.id === game.userId && cm.isDamageRoll ) {
-            damages.push(cm);
-            return false;
-        }
-    }
-
-    Hooks.on('preCreateChatMessage', PD);
-
-    const altUsage = null;
-    const ev = new KeyboardEvent('keydown', {'shiftKey': true});
-    const primaryMessage = await primary.variants[map].roll({ event:ev, altUsage });
-    const primaryDegreeOfSuccess = primaryMessage.degreeOfSuccess;
-    const secondaryMessage = await secondary.variants[map2].roll({ event:ev, altUsage });
-    const secondaryDegreeOfSuccess = secondaryMessage.degreeOfSuccess;
-
-    let pd,sd;
-    if ( primaryDegreeOfSuccess === 2 ) { pd = await primary.damage({event,options}); }
-    if ( primaryDegreeOfSuccess === 3 ) { pd = await primary.critical({event,options}); }
-    if ( secondaryDegreeOfSuccess === 2 ) { sd = await secondary.damage({event,options}); }
-    if ( secondaryDegreeOfSuccess === 3 ) { sd = await secondary.critical({event,options}); }
-
-    Hooks.off('preCreateChatMessage', PD);
-
-    if (damages.length === 0) {
-        ChatMessage.create({
-            type: CONST.CHAT_MESSAGE_TYPES.OOC,
-            content: "Both attacks missed"
-        });
-        return;
-    }
-
-
-    if ( (primaryDegreeOfSuccess <= 1 && secondaryDegreeOfSuccess >= 2) || (secondaryDegreeOfSuccess <= 1 && primaryDegreeOfSuccess >= 2)) {
-        ChatMessage.createDocuments(damages);
-        return;
-    }
-    await new Promise( (resolve) => {
-        setTimeout(resolve,0);
-    });
-
-    let flavor = '<strong>Flurry Of Blows Total Damage</strong>'
-    const color = (primaryDegreeOfSuccess || secondaryDegreeOfSuccess) === 2 ? `<span style="color:rgb(0, 0, 255)">Success</span>` : `<span style="color:rgb(0, 128, 0)">Critical Success</span>`
-    if (damages[0].flavor === damages[1].flavor) {
-        flavor += `<p>Same Weapon (${color})<hr>${damages[0].flavor}</p><hr>`;
-    } else {
-        flavor += `<hr>${damages[0].flavor}<hr>${damages[1].flavor}`;
-    }
-
-    const damageRolls = damages.map(a=>a.rolls).flat().map(a=>a.terms).flat().map(a=>a.rolls).flat();
-    const data = {};
-    for ( const dr of damageRolls ) {
-        if (dr.options.flavor in data) {
-            data[dr.options.flavor].push(dr.head.expression);
-        } else {
-            data[dr.options.flavor] = [dr.head.expression]
-        }
-    }
-    const formulas = [];
-    Object.keys(data).forEach(k=>{
-        console.log(k);
-        console.log(data[k]);
-         formulas.push(`(${data[k].join('+')})[${k}]`);
-    })
-
-    const rolls = [await new DamageRoll(formulas.join(',')).evaluate( {async: true} )];
-    const opts = damages[0].flags.pf2e.context.options.concat(damages[1].flags.pf2e.context.options);
-    await ChatMessage.create({
-        flags: {
-            pf2e: {
-                context: {
-                    options: [...new Set(opts)]
-                }
-            }
-        },
-        rolls,
-        type: CONST.CHAT_MESSAGE_TYPES.ROLL,
-        flavor,
-        speaker: ChatMessage.getSpeaker(),
-    });
+    combinedDamage("Flurry Of Blows", primary, secondary, options, map, map2);
 }
 
 Hooks.once("init", () => {